Grip and Autonomous push
Today we got Grip "working" still need to test to make sure it executes
properly... We also added some autonomous code so we can add a delay in
the front, we also implemented encoders to be used in autonomous.
Everything is untested still so no clue if this will function properly.