Encoders
We were able to find the correct Talon IDs that were connected to the
encoder wires. We are finally able to read encoder values! Notes: The
left encoder goes more and more negative the more you push the
joysticks, while the right goes more positive. Will try to fix tick
conversions in code.